requirement.txt
pyPDF2
streamlit
google-generativeai
langchain
chromadb
python-dotenv